+++ src/App.js
@@ -4,12 +4,12 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 import './App.css';
 import './fonts.css';
 
-import Home from './components/Home';
-import Dashboard from './components/Dashboard';
+import Home from './components/HomePage/Home';
+import Dashboard from './components/Dashboard/Dashboard';
 import LoginPage from './components/auth/LoginPage';
 import RegisterPage from './components/auth/RegisterPage';
 
-import MainView from './components/MainView';
+import MainView from './components/MainView/MainView';
 import DashboardLists from './components/list/Dashboard_Lists';
 import DashboardCampaigns from './components/campaign/Dashboard_Campaigns';
