+++ src/components/MainView/MainView.css
@@ -1,0 +1,5 @@
+.nav-bar-title {
+    font-family: 'Jaapokki';
+    color: black;
+    font-size: 10vw;
+}
