+++ src/components/MainView/MainView.js
@@ -2,9 +2,11 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import loginGate from './requires-login-gate';
+import loginGate from '../requires-login-gate';
 
-import { logout } from '../actions/loginActions';
+import { logout } from '../../actions/loginActions';
+
+import './MainView.css';
 
 export class MainView extends Component {
     
@@ -12,11 +14,13 @@ export class MainView extends Component {
 
         return(
             <div>
-                <div className='sideBar'>
-                    <h1><Link to='/'>Mailbox</Link></h1>
-                    <h3><Link to='/dashboard'>Dashboard</Link></h3>
-                    <h3><Link to='/dashboard/lists'>Lists</Link></h3>
-                    <h3><Link to='/dashboard/campaigns'>Campaigns</Link></h3>
+                <div className='nav-bar'>
+                    <Link to='/'><div className='nav-bar-title'>mailbox</div></Link>
+                    <div>
+                        <h3><Link to='/dashboard'>Dashboard</Link></h3>
+                        <h3><Link to='/dashboard/lists'>Lists</Link></h3>
+                        <h3><Link to='/dashboard/campaigns'>Campaigns</Link></h3>
+                    </div>
                 </div>
                 <button onClick={() => this.props.dispatch(logout())}>Log Out</button>
             </div>
