+++ src/components/auth/css/LoginPage.css
@@ -18,6 +18,7 @@
 .signup-link a {
     color: black;
 }
+
 /* desktop screens */
 @media (min-width: 600px) {
     .login-form-page {
